+++ tools/build-ruby-x86_64.sh
@@ -23,7 +23,7 @@ mkdir -p "$target_build"
 cd "$target_build"
 if [[ ! -f Makefile ]]; then
     CFLAGS="-O2 -ffreestanding -fno-stack-protector -fno-pie -mno-red-zone -U_FORTIFY_SOURCE -D_FORTIFY_SOURCE=0" \
-    "$source_dir/configure" --build=aarch64-unknown-linux-gnu --host=x86_64-unknown-none \
+    "$source_dir/configure" --build="$(gcc -dumpmachine)" --host=x86_64-unknown-none \
         --with-baseruby="$host_prefix/bin/ruby" --disable-shared \
         --disable-install-doc --disable-rubygems --disable-yjit --disable-zjit \
         --without-gmp --with-coroutine=amd64 --with-out-ext=all \